Added component structure
