+++ src/components/Quiz.css
@@ -1,0 +1,7 @@
+.button-selected {
+  background-color: green;
+}
+
+.button-unselected {
+  background-color: gray;
+}
