+++ src/components/ScoreMessage.js
@@ -1,0 +1,7 @@
+export default function ScoreMessage(score){
+  return (
+    <div className="ScoreMessage">
+      <p>Your score is: {score.score} out of 5</p>
+    </div>
+  );
+}
